split plot tests into 2 files
